// src/components/webgl-component/types.ts
export type ContentType = {
  name: string
  src: string
}

export type HeaderStyle = {
  position: string;
  zIndex: number;
  top: number;
  left: number;
  backgroundColor: string;
  width: number;
  height: number;
}

export type TitleContainerStyle = {
  position: string;
  zIndex: number;
  top: number;
  left: number;
  height: number;
}

export type TitleStyle = {
  color: string;
  fontSize: number;
}

export type LogoContainerStyle = {
  position: string;
  zIndex: number;
  top: number;
  left: number;
  width: number;
  height: number;
}

export type LogoStyle = {
  top: number;
  left: number;
  width: number;
  height: number;
}